fix: rename utils to requestValidation
