+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/MetaMask/template-snap-monorepo.git"
   },
   "source": {
-    "shasum": "wWVHOajslfBBkEDDlg162IwoG88t+frGKr0kOw01wj0=",
+    "shasum": "pBt7x9itaeOJLjDVhJxk4zRrYS9IAeM2cJI1YMxreu0=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
